extract multiline reformatter into separate object
